03_DeployAISolutions/05_MLOps/inference/requirements.txt
# python --version 3.12.9

fastapi==0.115.12
numpy==2.1.3
tensorflow==2.19.0
pillow==11.2.1
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0